refactor: change interface name
