+++ src/interfaces/index.tsx
@@ -1,4 +1,4 @@
 export * from './Session'
 export * from './User'
 export * from './ErrorMessage'
-export * from './IdToken'
+export * from './Token'
